# docs/api/openapi-v1-phase2.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Alter Earth API",
        "description": "A community platform for ecological conservation, sustainability advocacy, and using AI for environmental good",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/posts": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "List Posts",
                "description": "List posts with sorting and pagination.\n\nSupports multiple sort options:\n- hot: Reddit-style hot ranking (default)\n- top: Highest score in time period\n- new: Most recent first\n- controversial: Most debated\n- rising: Trending among new posts",
                "operationId": "list_posts_api_v1_posts_get",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/SortOption",
                            "default": "hot"
                        }
                    },
                    {
                        "name": "time_period",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "$ref": "#/components/schemas/TimePeriod",
                            "default": "day"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "default": 1,
                            "title": "Page"
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "default": 25,
                            "title": "Page Size"
                        }
                    },
                    {
                        "name": "community_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string",
                                    "format": "uuid"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Community Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Create User Submission",
                "description": "Create a new user submission (text or link post).\n\nRequires authentication.",
                "operationId": "create_user_submission_api_v1_posts_post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSubmissionCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/posts/{post_id}": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Get Post",
                "description": "Get a single post by ID with full details.",
                "operationId": "get_post_api_v1_posts__post_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Post Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/comments/post/{post_id}": {
            "get": {
                "tags": [
                    "comments"
                ],
                "summary": "Get Post Comments",
                "description": "Get all comments for a post as a threaded tree structure.\n\nReturns top-level comments with nested replies.",
                "operationId": "get_post_comments_api_v1_comments_post__post_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Post Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentTreeResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/votes/posts/{post_id}": {
            "post": {
                "tags": [
                    "votes"
                ],
                "summary": "Vote On Post",
                "description": "Vote on a post.\n\n- vote_value: 1 for upvote, -1 for downvote\n- If user already voted, updates the vote\n- Cannot vote on own posts\n- Vote weight is based on user's karma",
                "operationId": "vote_on_post_api_v1_votes_posts__post_id__post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Post Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VoteCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VoteResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "votes"
                ],
                "summary": "Remove Vote From Post",
                "description": "Remove vote from a post.\n\nReturns 204 even if vote didn't exist.",
                "operationId": "remove_vote_from_post_api_v1_votes_posts__post_id__delete",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Post Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CommentTreeResponse": {
                "properties": {
                    "post_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Post Id"
                    },
                    "total_comments": {
                        "type": "integer",
                        "title": "Total Comments"
                    },
                    "top_level_count": {
                        "type": "integer",
                        "title": "Top Level Count"
                    },
                    "comments": {
                        "items": {
                            "$ref": "#/components/schemas/CommentWithReplies"
                        },
                        "type": "array",
                        "title": "Comments"
                    }
                },
                "type": "object",
                "required": [
                    "post_id",
                    "total_comments",
                    "top_level_count",
                    "comments"
                ],
                "title": "CommentTreeResponse",
                "description": "Schema for complete comment tree on a post.",
                "example": {
                    "comments": [],
                    "post_id": "550e8400-e29b-41d4-a716-446655440000",
                    "top_level_count": 5,
                    "total_comments": 15
                }
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PostListResponse": {
                "properties": {
                    "posts": {
                        "items": {
                            "$ref": "#/components/schemas/PostListItem"
                        },
                        "type": "array",
                        "title": "Posts"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    },
                    "page": {
                        "type": "integer",
                        "title": "Page"
                    },
                    "page_size": {
                        "type": "integer",
                        "title": "Page Size"
                    },
                    "has_more": {
                        "type": "boolean",
                        "title": "Has More"
                    }
                },
                "type": "object",
                "required": [
                    "posts",
                    "total",
                    "page",
                    "page_size",
                    "has_more"
                ],
                "title": "PostListResponse",
                "description": "Paginated list of posts."
            },
            "PostResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "entity_type": {
                        "$ref": "#/components/schemas/EntityTypeEnum"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "User Id"
                    },
                    "community_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Community Id"
                    },
                    "upvotes": {
                        "type": "integer",
                        "title": "Upvotes"
                    },
                    "downvotes": {
                        "type": "integer",
                        "title": "Downvotes"
                    },
                    "score": {
                        "type": "integer",
                        "title": "Score"
                    },
                    "hot_score": {
                        "type": "number",
                        "title": "Hot Score"
                    },
                    "view_count": {
                        "type": "integer",
                        "title": "View Count"
                    },
                    "comment_count": {
                        "type": "integer",
                        "title": "Comment Count"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    },
                    "automated_article": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/AutomatedArticleResponse"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "user_submission": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/UserSubmissionResponse"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "user_nickname": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User Nickname"
                    },
                    "user_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User Type"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "title",
                    "entity_type",
                    "user_id",
                    "community_id",
                    "upvotes",
                    "downvotes",
                    "score",
                    "hot_score",
                    "view_count",
                    "comment_count",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "title": "PostResponse",
                "description": "Schema for post in API responses."
            },
            "SortOption": {
                "type": "string",
                "enum": [
                    "hot",
                    "top",
                    "new",
                    "controversial",
                    "rising"
                ],
                "title": "SortOption",
                "description": "Post sorting options."
            },
...
            "TimePeriod": {
                "type": "string",
                "enum": [
                    "hour",
                    "day",
                    "week",
                    "month",
                    "year",
                    "all"
                ],
                "title": "TimePeriod",
                "description": "Time period for 'top' sorting."
            },
...
            "UserSubmissionCreate": {
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 500,
                        "minLength": 1,
                        "title": "Title"
                    },
                    "submission_type": {
                        "$ref": "#/components/schemas/SubmissionTypeEnum"
                    },
                    "content": {
                        "anyOf": [
                            {
                                "type": "string",
                                "minLength": 1
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Content"
                    },
                    "url": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 1000
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Url"
                    }
                },
                "type": "object",
                "required": [
                    "title",
                    "submission_type"
                ],
                "title": "UserSubmissionCreate",
                "description": "Schema for creating a user submission.",
                "examples": [
                    {
                        "content": "Ocean ecosystems are facing unprecedented threats...",
                        "submission_type": "text",
                        "title": "Why we need to protect ocean ecosystems"
                    },
                    {
                        "submission_type": "link",
                        "title": "New study on coral reef restoration",
                        "url": "https://example.com/coral-study"
                    }
                ]
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "VoteCreate": {
                "properties": {
                    "vote_value": {
                        "type": "integer",
                        "enum": [
                            1,
                            -1
                        ],
                        "title": "Vote Value",
                        "description": "1 for upvote, -1 for downvote"
                    }
                },
                "type": "object",
                "required": [
                    "vote_value"
                ],
                "title": "VoteCreate",
                "description": "Schema for creating or updating a vote.",
                "examples": [
                    {
                        "vote_value": 1
                    },
                    {
                        "vote_value": -1
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}